* commit '8bf81bd2e2007bb9fcbd7934fe4cf9aeec24a08f':
  PHP error when opening 'Add from files' dialog

+++ php/Files/Core/Util/class.stringutil.php
@@ -23,7 +23,7 @@ class StringUtil
 	static function human_filesize($bytes, $decimals = 2)
 	{
 		$sz = ' KMGTP';
-		$factor = floor((strlen($bytes) - 1) / 3);
+		$factor = (int)floor((strlen($bytes) - 1) / 3);
 		return sprintf("%.{$decimals}f", $bytes / pow(1024, $factor)) . " " . @$sz[$factor] . "B";
 	}
